fixed routing for startup/student apps cuz *jiya* dun fuc up on emails >:| also added mark pfp
