--- src/app/apply-students/success/page.module.css
@@ -1,169 +1,0 @@
-/* Success Page Styles */
-.successPage {
-  background: #000000;
-  min-height: 100vh;
-  position: relative;
-  overflow: hidden;
-}
-
-.successPage::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background:
-    radial-gradient(circle at 20% 20%, rgba(66, 104, 255, 0.1) 0%, transparent 50%),
-    radial-gradient(circle at 80% 80%, rgba(66, 104, 255, 0.1) 0%, transparent 50%);
-  pointer-events: none;
-}
-
-.successContent {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 6rem 2rem 4rem;
-  text-align: center;
-  position: relative;
-  z-index: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  min-height: calc(100vh - 12rem);
-}
-
-/* Header Navigation */
-.headerWithNav {
-  position: absolute;
-  top: max(16px, env(safe-area-inset-top));
-  left: max(16px, env(safe-area-inset-left));
-  z-index: 2;
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.returnButton {
-  background: rgba(255, 255, 255, 0.05);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  border-radius: 8px;
-  padding: 0.75rem 1.5rem;
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.8);
-  cursor: pointer;
-  transition: all 0.3s ease;
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.returnButton:hover {
-  background: rgba(255, 255, 255, 0.08);
-  border-color: rgba(66, 104, 255, 0.5);
-  color: #ffffff;
-  transform: translateX(-2px);
-}
-
-.returnButton:focus-visible {
-  outline: 2px solid rgba(66, 104, 255, 0.9);
-  outline-offset: 3px;
-}
-
-/* Title and Subtitle */
-.title {
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: clamp(2.5rem, 5vw, 3.5rem);
-  font-weight: 700;
-  color: #ffffff;
-  margin-bottom: 1.5rem;
-  letter-spacing: -0.02em;
-  animation: fadeInUp 0.6s ease-out 0.2s both;
-}
-
-.subtitle {
-  font-size: clamp(1.125rem, 2vw, 1.25rem);
-  color: rgba(255, 255, 255, 0.8);
-  line-height: 1.6;
-  margin-bottom: 4rem;
-  max-width: 600px;
-  margin-left: auto;
-  margin-right: auto;
-  animation: fadeInUp 0.6s ease-out 0.4s both;
-}
-
-/* Contact Info */
-.contactInfo {
-  background: rgba(66, 104, 255, 0.05);
-  border: 1px solid rgba(66, 104, 255, 0.2);
-  border-radius: 16px;
-  padding: 2rem;
-  margin: 0 auto 3rem;
-  max-width: 650px;
-  animation: fadeInUp 0.6s ease-out 0.6s both;
-}
-
-.contactTitle {
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: #ffffff;
-  margin-bottom: 1rem;
-}
-
-.contactText {
-  color: rgba(255, 255, 255, 0.8);
-  line-height: 1.6;
-  margin: 0;
-}
-
-.emailLink {
-  color: #4268ff;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.3s ease;
-}
-
-.emailLink:hover {
-  color: #667eea;
-  text-decoration: underline;
-}
-
-/* Animations */
-@keyframes fadeInUp {
-  from {
-    opacity: 0;
-    transform: translateY(30px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .successContent {
-    padding: 4rem 1rem 2rem;
-  }
-  
-  .contactInfo {
-    padding: 1.5rem;
-    margin: 0 auto 2rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .title {
-    font-size: 2rem;
-  }
-  
-  .subtitle {
-    font-size: 1rem;
-  }
-  
-  .contactTitle {
-    font-size: 1.25rem;
-  }
-}
