+++ src/app/apply/success/page.module.css
@@ -71,31 +71,6 @@
   outline-offset: 3px;
 }
 
-/* Success Icon */
-.successIcon {
-  font-size: 4rem;
-  margin-bottom: 2rem;
-  animation: bounceIn 0.8s ease-out;
-}
-
-@keyframes bounceIn {
-  0% {
-    opacity: 0;
-    transform: scale(0.3);
-  }
-  50% {
-    opacity: 1;
-    transform: scale(1.05);
-  }
-  70% {
-    transform: scale(0.9);
-  }
-  100% {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
 /* Title and Subtitle */
 .title {
   font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
@@ -118,67 +93,6 @@
   animation: fadeInUp 0.6s ease-out 0.4s both;
 }
 
-/* Next Steps Section */
-.nextSteps {
-  background: rgba(255, 255, 255, 0.03);
-  backdrop-filter: blur(20px);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 20px;
-  padding: 3rem;
-  margin-bottom: 3rem;
-  text-align: left;
-  animation: fadeInUp 0.6s ease-out 0.6s both;
-}
-
-.nextStepsTitle {
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 1.75rem;
-  font-weight: 600;
-  color: #ffffff;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.steps {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.step {
-  display: flex;
-  align-items: flex-start;
-  gap: 1.5rem;
-}
-
-.stepNumber {
-  background: linear-gradient(135deg, #4268ff, #667eea);
-  color: #ffffff;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: 600;
-  font-size: 1.125rem;
-  flex-shrink: 0;
-}
-
-.stepContent h3 {
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #ffffff;
-  margin-bottom: 0.5rem;
-}
-
-.stepContent p {
-  color: rgba(255, 255, 255, 0.7);
-  line-height: 1.6;
-  margin: 0;
-}
-
 /* Contact Info */
 .contactInfo {
   background: rgba(66, 104, 255, 0.05);
@@ -187,7 +101,7 @@
   padding: 2rem;
   margin: 0 auto 3rem;
   max-width: 650px;
-  animation: fadeInUp 0.6s ease-out 0.8s both;
+  animation: fadeInUp 0.6s ease-out 0.6s both;
 }
 
 .contactTitle {
@@ -216,93 +130,6 @@
   text-decoration: underline;
 }
 
-/* Action Buttons */
-.actions {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  margin-bottom: 3rem;
-  animation: fadeInUp 0.6s ease-out 1s both;
-}
-
-.primaryButton {
-  background: linear-gradient(135deg, #4268ff, #667eea);
-  border: none;
-  border-radius: 12px;
-  padding: 1rem 2rem;
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 1rem;
-  font-weight: 600;
-  color: #ffffff;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.primaryButton:hover {
-  background: linear-gradient(135deg, #3557e6, #5a6fd8);
-  transform: translateY(-2px);
-  box-shadow: 0 8px 25px rgba(66, 104, 255, 0.3);
-}
-
-.secondaryButton {
-  background: rgba(255, 255, 255, 0.05);
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  border-radius: 12px;
-  padding: 1rem 2rem;
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 1rem;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.8);
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.secondaryButton:hover {
-  background: rgba(255, 255, 255, 0.08);
-  border-color: rgba(255, 255, 255, 0.3);
-  color: #ffffff;
-  transform: translateY(-2px);
-}
-
-/* Social Links */
-.socialLinks {
-  animation: fadeInUp 0.6s ease-out 1.2s both;
-}
-
-.socialText {
-  color: rgba(255, 255, 255, 0.6);
-  font-size: 0.875rem;
-  margin-bottom: 1rem;
-}
-
-.socialButtons {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-}
-
-.socialLink {
-  background: rgba(255, 255, 255, 0.05);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 8px;
-  padding: 0.75rem 1.5rem;
-  font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.8);
-  text-decoration: none;
-  transition: all 0.3s ease;
-}
-
-.socialLink:hover {
-  background: rgba(255, 255, 255, 0.08);
-  border-color: rgba(66, 104, 255, 0.3);
-  color: #ffffff;
-  transform: translateY(-1px);
-}
-
 /* Animations */
 @keyframes fadeInUp {
   from {
@@ -321,40 +148,9 @@
     padding: 4rem 1rem 2rem;
   }
   
-  .nextSteps {
-    padding: 2rem;
-  }
-  
-  .actions {
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .primaryButton,
-  .secondaryButton {
-    width: 100%;
-    max-width: 300px;
-  }
-  
-  .step {
-    flex-direction: column;
-    text-align: center;
-    gap: 1rem;
-  }
-  
-  .stepNumber {
-    align-self: center;
-  }
-  
-  .socialButtons {
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .socialLink {
-    width: 100%;
-    max-width: 200px;
-    text-align: center;
+  .contactInfo {
+    padding: 1.5rem;
+    margin: 0 auto 2rem;
   }
 }
 
@@ -367,11 +163,7 @@
     font-size: 1rem;
   }
   
-  .nextStepsTitle {
-    font-size: 1.5rem;
-  }
-  
   .contactTitle {
     font-size: 1.25rem;
   }
-} 
+}
