+++ src/app/components/index.html
@@ -41,8 +41,8 @@
           </ul>
         </nav>
         <div class="btn-actions">
-          <a href="/apply-students" class="apply-btn">Students</a>
-          <a href="/apply" class="apply-btn">Startups</a>
+          <a href="/apply" class="apply-btn">Students</a>
+          <a href="/apply-startups" class="apply-btn">Startups</a>
         </div>
       </div>
     </header>
@@ -59,11 +59,11 @@
           <p class="hero-subtitle">Healthcare & Biotech Startup Incubator</p>
 
           <div class="hero-actions">
-            <a href="/apply-students" target="_self" class="action-card">
+            <a href="/apply" target="_self" class="action-card">
               <i class="bi bi-person-plus"></i>
               <h3>Apply for Students</h3>
             </a>
-            <a href="/apply" target="_self" class="action-card">
+            <a href="/apply-startups" target="_self" class="action-card">
               <i class="bi bi-building"></i>
               <h3>Apply for Startups</h3>
             </a>
@@ -473,7 +473,9 @@
 
           <div class="team-member">
             <div class="member-image">
-              <div class="member-placeholder">MA</div>
+              <div class="member-placeholder">
+                <img src="assets/img/team/mark.jpg" alt="" />
+              </div>
               <div class="social-links">
                 <a href="#"><i class="bi bi-twitter-x"></i></a>
                 <a href="#"><i class="bi bi-linkedin"></i></a>
@@ -750,8 +752,8 @@
           <div class="spacer-content">
             <h3>Ready to Join Yale Helix?</h3>
             <div class="application-buttons">
-              <a href="/apply" target="_self" class="app-button startup">Startup Application</a>
-              <a href="/apply-students" target="_self" class="app-button student">Student Application</a>
+              <a href="/apply-startups" target="_self" class="app-button startup">Startup Application</a>
+              <a href="/apply" target="_self" class="app-button student">Student Application</a>
             </div>
           </div>
         </div>
